+++ MVVM-Fishing/Models/Structs and Classes/Inventory/Bait.swift
@@ -19,9 +19,9 @@ class Bait: Object, ItemType {
     @objc dynamic var imagePath: String = ""
     @objc dynamic var name: String = ""
     /// Represents the amount of bait you get when you buy one item.
-    @objc dynamic var value: Float = 0.0
+    @objc dynamic var value: Int = 0
     
-    convenience init(premium: Bool, price: Int, imagePath: String, name: String, value: Float) {
+    convenience init(premium: Bool, price: Int, imagePath: String, name: String, value: Int) {
         self.init()
         self.premium = premium
         self.price = price
