+++ MVVM-Fishing/Models/Structs and Classes/Inventory/Hook.swift
@@ -16,9 +16,9 @@ class Hook: Object, ItemType {
     @objc dynamic var price: Int = 0
     @objc dynamic var imagePath: String = ""
     @objc dynamic var name: String = ""
-    @objc dynamic var value: Float = 0.0
+    @objc dynamic var value: Int = 0
     
-    convenience init(premium: Bool, price: Int, imagePath: String, name: String, value: Float) {
+    convenience init(premium: Bool, price: Int, imagePath: String, name: String, value: Int) {
         self.init()
         self.premium = premium
         self.price = price
